Move Dato transforms into publish content script
Doing this will make it easier to publish content into other sources
such as Algolia.
